+++ Event.h
@@ -7,19 +7,19 @@ class Event
    friend ostream& operator<<(ostream &outStream, const Event &e);
 
 public:
-   Event(string eventType, int time, int serviceTime, int extraTime);
+   Event(string eventType, double time, double serviceTime, double extraTime);
    string getEventType() const;
-   int getTime() const;
-   int getServiceLength() const;
-   int getExtraTime() const;
-   int getTotalServiceTime() const;
+   double getTime() const;
+   double getServiceLength() const;
+   double getExtraTime() const;
+   double getTotalServiceTime() const;
    bool operator<(const Event &e1) const;
    bool operator>(const Event &e1) const;
 
 private:
    string eventType_;
-   int time_;
-   int serviceLength_;
-   int extraTime_;
-   int totalServiceTime_;
+   double time_;
+   double serviceLength_;
+   double extraTime_;
+   double totalServiceTime_;
 };
